Move incoming message parsing to parse.ts
